pyproject.toml
[tool.poetry]
name = "opportunity-backend"
version = "0.1.0"
description = "A platform to help teens find opportunities for them"
authors = ["Dan Percic <danpercic86@gmail.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^5.4.1"
psycopg2 = "^2.8.6"
djangorestframework = "^3.12.2"
drf-spectacular = "^0.13.1"
Pillow = "^8.1.0"
django-cors-headers = "^3.5.0"
django-pwa = "^1.0.10"
django-admin-thumbnails = "^0.2.5"
django-admin-sortable2 = "0.7.7"
django-model-utils = "^4.1.1"
djangorestframework-simplejwt = "^4.6.0"
django-jazzmin = "^2.4.4"
Django = "^3.1.6"

[tool.poetry.dev-dependencies]
pylint = "2.6.0"
black = "^20.8b1"
pycodestyle = "^2.6.0"
django-querycount = "^0.7.0"
django-debug-toolbar = "^3.2"
nplusone = "^1.0.0"
django-extra-checks = "^0.8.0"

[tool.black]
target-version = ['py38']
exclude = '''
/(
    \.git
  | \__pycache__
  | \venv
  | \.venv
  | env
  | \.env
)/
'''

[tool.pylint.messages_control]
disable = """
    C0330,
    C0326,
    missing-class-docstring,
    missing-module-docstring,
    too-few-public-methods,
    unused-wildcard-import,
    wildcard-import,
    unsubscriptable-object,
    missing-function-docstring,
    ungrouped-imports,
    no-member,
    too-many-ancestors,
    import-outside-toplevel,
    protected-access,
    unused-argument,
    wrong-import-position
"""

[tool.pylint.master]
ignore = """
    migrations
"""

[tool.pylint.format]
max-line-length = "88"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
